Pushing the changes for product delete, find by sku , update restapi

+++ src/main/java/com/products/service/elevatemartproductsservice/services/ProductService.java
@@ -8,7 +8,7 @@ public interface ProductService {
 
     void createProduct(Product product);
     void updateProduct(Product product);
-    void deleteProduct(Integer productId);
+    void deleteProduct(Long productId);
 
     void updateProductSku(Long productId, String updatedSku);
     void updateProductQuantities(String sku, Long quantities);
